+++ app/controllers/projects/milestones_controller.rb
@@ -42,6 +42,7 @@ class Projects::MilestonesController < Projects::ApplicationController
   end
 
   def show
+    @burndown = Burndown.new(@milestone)
   end
 
   def create