+++ config/application.rb
@@ -153,6 +153,7 @@ module Gitlab
 
     # This is needed for gitlab-shell
     ENV['GITLAB_PATH_OUTSIDE_HOOK'] = ENV['PATH']
+    ENV['GIT_TERMINAL_PROMPT'] = '0'
 
     # Gitlab Geo Middleware support
     config.middleware.insert_after ActionDispatch::Flash, 'Gitlab::Middleware::ReadonlyGeo'