+++ config/webpack.config.js
@@ -23,6 +23,7 @@ var config = {
     main:                 './main.js',
     blob:                 './blob_edit/blob_bundle.js',
     boards:               './boards/boards_bundle.js',
+    burndown_chart:       './burndown_chart/index.js',
     cycle_analytics:      './cycle_analytics/cycle_analytics_bundle.js',
     commit_pipelines:     './commit/pipelines/pipelines_bundle.js',
     diff_notes:           './diff_notes/diff_notes_bundle.js',
@@ -39,6 +40,8 @@ var config = {
     monitoring:           './monitoring/monitoring_bundle.js',
     network:              './network/network_bundle.js',
     notebook_viewer:      './blob/notebook_viewer.js',
+    sketch_viewer:        './blob/sketch_viewer.js',
+    pdf_viewer:           './blob/pdf_viewer.js',
     profile:              './profile/profile_bundle.js',
     protected_branches:   './protected_branches/protected_branches_bundle.js',
     snippet:              './snippet/snippet_bundle.js',
@@ -66,7 +69,11 @@ var config = {
       {
         test: /\.svg$/,
         use: 'raw-loader'
-      }
+      }, {
+        test: /\.(worker.js|pdf)$/,
+        exclude: /node_modules/,
+        loader: 'file-loader',
+      },
     ]
   },
 
@@ -108,6 +115,7 @@ var config = {
         'issuable',
         'merge_conflicts',
         'notebook_viewer',
+        'pdf_viewer',
         'vue_pipelines',
         'mr_widget_ee',
       ],
@@ -123,6 +131,7 @@ var config = {
         'graphs',
         'users',
         'monitoring',
+        'burndown_chart',
       ],
     }),