+++ spec/features/merge_requests/create_new_mr_spec.rb
@@ -15,7 +15,7 @@ feature 'Create New Merge Request', feature: true, js: true do
   it 'selects the source branch sha when a tag with the same name exists' do
     visit namespace_project_merge_requests_path(project.namespace, project)
 
-    click_link 'New Merge Request'
+    click_link 'New merge request'
     expect(page).to have_content('Source branch')
     expect(page).to have_content('Target branch')
 
@@ -27,8 +27,8 @@ feature 'Create New Merge Request', feature: true, js: true do
 
   it 'selects the target branch sha when a tag with the same name exists' do
     visit namespace_project_merge_requests_path(project.namespace, project)
-    
-    click_link 'New Merge Request'
+
+    click_link 'New merge request'
 
     expect(page).to have_content('Source branch')
     expect(page).to have_content('Target branch')
@@ -42,7 +42,7 @@ feature 'Create New Merge Request', feature: true, js: true do
   it 'generates a diff for an orphaned branch' do
     visit namespace_project_merge_requests_path(project.namespace, project)
 
-    click_link 'New Merge Request'
+    page.has_link?('New Merge Request') ? click_link("New Merge Request") : click_link('New merge request')
     expect(page).to have_content('Source branch')
     expect(page).to have_content('Target branch')