+++ spec/javascripts/environments/environment_spec.js
@@ -279,8 +279,9 @@ describe('Environment', () => {
           setTimeout(() => {
             expect(component.$el.querySelectorAll('.js-child-row').length).toEqual(1);
             expect(component.$el.querySelector('td.text-center > a.btn').textContent).toContain('Show all');
-            done();
+
             Vue.http.interceptors = _.without(Vue.http.interceptors, folderInterceptor);
+            done();
           });
         });
       });